10/10

+++ JavaProjects/projeto_sistema_manutencao_somativa/src/main/java/com/example/api/FalhaAPI.java
@@ -1,12 +1,10 @@
 package com.example.api;
 
+import org.json.JSONArray;
+import org.json.JSONObject;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
-
-import org.json.JSONArray;
-import org.json.JSONObject;
-
 import com.example.models.Falha;
 
 public class FalhaAPI {
@@ -14,14 +12,15 @@ public class FalhaAPI {
         String json = ApiConnection.getData("falhas");
         List<Falha> falhas = new ArrayList<>();
 
-
         if (json != null) {
             JSONArray jsonArray = new JSONArray(json);
             for (int i = 0; i < jsonArray.length(); i++) {
                 JSONObject jsonObject = jsonArray.getJSONObject(i);
+                
+                
                 Falha falha = new Falha(
                     jsonObject.getString("id"),
-                    jsonObject.getInt("maquinaId"),
+                    jsonObject.getLong("maquinaId"),
                     LocalDate.parse(jsonObject.getString("data")),
                     jsonObject.getString("problema"),
                     jsonObject.getString("prioridade"),
@@ -32,5 +31,4 @@ public class FalhaAPI {
         }
         return falhas;
     }
-
 }
